Updates to handle SWIFT HDF5 input

+++ src/hdfitems.h
@@ -663,7 +663,7 @@ struct HDF_Group_Names {
 struct HDF_Header {
 
     double      BoxSize;
-    unsigned long int npart[NHDFTYPE];
+    unsigned long long npart[NHDFTYPE];
     unsigned int npartTotal[NHDFTYPE];
     unsigned int npartTotalHW[NHDFTYPE];
     double      mass[NHDFTYPE];
